add updated example script files
These are the atlantisom example scripts I developed based on the previous scripts. Basically, I did some reorganizing and added more comments to provide additional information and guidance. Also, I have a version of the script with 2 fleets (1 fishery and 1 survey) and a version of the script with 6 fleets (3 fisheries and 3 surveys).